Refactor Utils file. Split format functions into new file  (#1480)
* Refactor format functions to new utils file

* Fix bad merge

* Fix merge

* Fix history

* Fix tests

* Resolve potential NPE for e2e tests

* Fix test
